+++ ckan/logic/action/delete.py
@@ -7,13 +7,11 @@ import logging
 import sqlalchemy as sqla
 import six
 
-import ckan.lib.jobs as jobs
+from ckan.lib import api_token, dictization, jobs, uploader
 import ckan.logic
 import ckan.logic.action
 import ckan.plugins as plugins
-import ckan.lib.dictization as dictization
 import ckan.lib.dictization.model_dictize as model_dictize
-import ckan.lib.api_token as api_token
 from ckan import authz
 
 from ckan.common import _
@@ -195,6 +193,15 @@ def resource_delete(context, data_dict):
     for plugin in plugins.PluginImplementations(plugins.IResourceController):
         plugin.before_delete(context, data_dict,
                              pkg_dict.get('resources', []))
+
+    # Delete file if it was uploaded
+    if entity.get('url_type') == 'upload':
+        upload = uploader.get_resource_uploader(entity)
+        # Don't break if old plugin/interface is found
+        if hasattr(upload, "delete"):
+            upload.delete(id)
+        else:
+            logging.warning("%s does not have delete function, could not cleanup: %s", type(upload).__name__, id)
 
     pkg_dict = _get_action('package_show')(context, {'id': package_id})