+++ ckan/tests/controllers/test_package.py
@@ -752,6 +752,31 @@ class TestPackageRead(object):
             status=404,
         )
 
+    # Test the 'reveal_private_datasets' flag
+
+    @pytest.mark.ckan_config("ckan.auth.reveal_private_datasets", "True")
+    def test_anonymous_users_cannot_read_private_datasets(self, app):
+        organization = factories.Organization()
+        dataset = factories.Dataset(owner_org=organization["id"], private=True)
+        response = app.get(
+            url_for("dataset.read", id=dataset["name"]),
+            follow_redirects=False
+        )
+        assert 302 == response.status_code
+        assert '/login' in response.headers[u"Location"]
+
+    @pytest.mark.ckan_config("ckan.auth.reveal_private_datasets", "True")
+    def test_user_not_in_organization_cannot_read_private_datasets(self, app):
+        user = factories.User()
+        organization = factories.Organization()
+        dataset = factories.Dataset(owner_org=organization["id"], private=True)
+        response = app.get(
+            url_for("dataset.read", id=dataset["name"]),
+            extra_environ={"REMOTE_USER": six.ensure_str(user["name"])},
+            status=403,
+        )
+        assert 403 == response.status_code
+
 
 @pytest.mark.usefixtures("clean_db", "with_request_context")
 class TestPackageDelete(object):
@@ -1308,6 +1333,36 @@ class TestResourceRead(object):
             url_for("dataset.read", id=dataset["name"]), status=404
         )
         assert 404 == response.status_code
+
+    # Test the 'reveal_private_datasets' flag
+
+    @pytest.mark.ckan_config("ckan.auth.reveal_private_datasets", "True")
+    def test_user_not_in_organization_cannot_read_resources_in_private_dataset(self, app):
+        user = factories.User()
+        env = {"REMOTE_USER": six.ensure_str(user["name"])}
+        organization = factories.Organization()
+        dataset = factories.Dataset(owner_org=organization["id"], private=True)
+        resource = factories.Resource(package_id=dataset["id"])
+
+        url = url_for(
+            "{}_resource.read".format(dataset["type"]),
+            id=dataset["id"], resource_id=resource["id"]
+        )
+
+        response = app.get(url, status=403, extra_environ=env)
+
+    @pytest.mark.ckan_config("ckan.auth.reveal_private_datasets", "True")
+    def test_anonymous_users_cannot_read_resources_in_private_datasets(self, app):
+        organization = factories.Organization()
+        dataset = factories.Dataset(owner_org=organization["id"], private=True)
+        resource = factories.Resource(package_id=dataset["id"])
+        response = app.get(
+            url_for("{}_resource.read".format(dataset["type"]),
+                    id=dataset["id"], resource_id=resource["id"]),
+            follow_redirects=False
+        )
+        assert 302 == response.status_code
+        assert '/login' in response.headers[u"Location"]
 
 
 @pytest.mark.usefixtures("clean_db", "with_request_context")