+++ ckan/tests/logic/action/test_get.py
@@ -4,8 +4,9 @@ import datetime
 import re
 
 import copy
+import mock
 import pytest
-from six import text_type
+from six import PY3, text_type
 from six.moves import xrange
 
 from ckan import model
@@ -16,7 +17,17 @@ import ckan.tests.factories as factories
 import ckan.tests.helpers as helpers
 import ckan.plugins.toolkit as tk
 from ckan import __version__
+from ckan.lib import uploader as ckan_uploader
 from ckan.lib.search.common import SearchError
+from pyfakefs import fake_filesystem
+
+fs = fake_filesystem.FakeFilesystem()
+fake_os = fake_filesystem.FakeOsModule(fs)
+fake_open = fake_filesystem.FakeFileOpen(fs)
+if PY3:
+    mock_open_path = 'builtins.open'
+else:
+    mock_open_path = '__builtin__.open'
 
 
 @pytest.mark.usefixtures("clean_db", "with_request_context")
@@ -2446,6 +2457,58 @@ class TestShowResourceView(object):
             helpers.call_action("resource_view_show", id="does_not_exist")
 
 
+def mock_open_if_open_fails(*args, **kwargs):
+    try:
+        return real_open(*args, **kwargs)
+    except (OSError, IOError):
+        return fake_open(*args, **kwargs)
+
+
+class ShowResourceFileMetadata(object):
+
+    @helpers.change_config('ckan.storage_path', '/doesnt_exist')
+    @mock.patch(mock_open_path, side_effect=mock_open_if_open_fails)
+    @mock.patch.object(ckan_uploader, 'os', fake_os)
+    @mock.patch.object(ckan_uploader, '_storage_path', new='/doesnt_exist')
+    def test_resource_file_metadata_show_meta_data_returned(self, _):
+        user = factories.User()
+        pkg = factories.Dataset(creator_user_id=user['id'])
+
+        url = url_for(
+            controller='api',
+            action='action',
+            logic_function='resource_create', ver='/3')
+        env = {'REMOTE_USER': user['name'].encode('ascii')}
+        postparams = {
+            'name': 'test-flask-upload',
+            'package_id': pkg['id']
+        }
+        upload_content = 'test-content,and2'
+        upload_info = ('upload', 'test-upload.csv', upload_content)
+        app = self._get_test_app()
+        resp = app.post(
+            url, params=postparams,
+            upload_files=[upload_info],
+            extra_environ=env
+        )
+        result = resp.json['result']
+        eq('upload', result['url_type'])
+        eq(len(upload_content), result['size'])
+
+        metaresult = helpers.call_action('resource_file_metadata_show', id=result['id'])
+        eq(len(upload_content), metaresult['size'])
+        eq('text/csv', metaresult['content_type'])
+        eq('', metaresult['hash'])
+
+    def test_resource_file_metadata_show_id_missing(self):
+        with pytest.raises(logic.ValidationError):
+            helpers.call_action('resource_file_metadata_show')
+
+    def test_resource_file_metadata_show_id_not_found(self):
+        with pytest.raises(logic.NotFound):
+            helpers.call_action('resource_file_metadata_show', id='does_not_exist')
+
+
 class TestGetHelpShow(object):
     def test_help_show_basic(self):
 
@@ -2819,6 +2882,20 @@ class TestStatusShow(object):
 
         assert type(status[u"extensions"]) == list
         assert status[u"extensions"] == [u"stats"]
+
+    @helpers.change_config('ckan.hide_version', 'True')
+    def test_status_show_hiding_version(self):
+
+        status = helpers.call_action(u'status_show')
+
+        assert 'ckan_version' not in status, "Should have skipped CKAN version"
+        assert status[u'site_url'] == u'http://test.ckan.net'
+        assert status[u'site_title'] == u'CKAN'
+        assert status[u'site_description'] == u''
+        assert status[u'locale_default'] == u'en'
+
+        assert type(status[u'extensions']) == list
+        assert status[u'extensions'] == [u'stats']
 
 
 class TestJobList(helpers.FunctionalRQTestBase):